requirements.txt
# Data Scraping & DB Connection
telethon
psycopg2-binary

# Data Transformation
dbt-postgres

# Data Enrichment
ultralytics
# torch  <-- COMMENT OUT OR DELETE THIS
# torchvision <-- COMMENT OUT OR DELETE THIS
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1+cpu
torchvision==0.18.1+cpu


# API
fastapi
uvicorn[standard]

# Orchestration
dagster
dagster-webserver
dagster-postgres

# Utilities
python-dotenv